Fixed BigQuery Multi Sink Null Pointer Exception due to not passing in a table
